Split Profile controller
