+++ src/Routes.php
@@ -38,9 +38,9 @@ class Routes
         // "My" section
         $router->addRoute('get', '/my/profile', 'my/Profile#show', 'profile');
         $router->addRoute('post', '/my/profile', 'my/Profile#update', 'update profile');
-        $router->addRoute('post', '/my/profile/avatar', 'my/Profile#updateAvatar', 'update avatar');
+        $router->addRoute('post', '/my/profile/avatar', 'my/Avatar#update', 'update avatar');
 
-        $router->addRoute('get', '/my/info.json', 'my/Profile#info', 'profile info');
+        $router->addRoute('get', '/my/info.json', 'my/Info#show', 'profile info');
 
         $router->addRoute('get', '/my/security', 'my/Security#show', 'security');
         $router->addRoute('post', '/my/security', 'my/Security#update', 'update security');
